+++ core-ng-mongo/src/test/java/core/framework/mongo/module/MongoConfigTest.java
@@ -1,8 +1,11 @@
 package core.framework.mongo.module;
 
+import com.mongodb.ConnectionString;
+import core.framework.internal.module.ReadinessProbe;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
@@ -19,17 +22,27 @@ class MongoConfigTest {
     @Test
     void validate() {
         assertThatThrownBy(() -> config.validate())
-                .hasMessageContaining("mongo uri must be configured");
+            .hasMessageContaining("mongo uri must be configured");
 
         config.uri = "mongodb://uri/db";
 
         assertThatThrownBy(() -> config.validate())
-                .hasMessageContaining("no collection/view added");
+            .hasMessageContaining("no collection/view added");
     }
 
     @Test
     void uri() {
         assertThatThrownBy(() -> config.uri("mongodb://localhost"))
-                .hasMessageContaining("uri must have database");
+            .hasMessageContaining("uri must have database");
+    }
+
+    @Test
+    void addProbe() {
+        var probe = new ReadinessProbe();
+        config.addProbe(probe, new ConnectionString("mongodb+srv://server.example.com/db"));
+        assertThat(probe.hostURIs).isEmpty();
+
+        config.addProbe(probe, new ConnectionString("mongodb://server.example.com/db"));
+        assertThat(probe.hostURIs).hasSize(1);
     }
 }